Create README.md
Assignment 2 of Mobile App development where the application has the layout of a signup form. If fields are left blank, the related error message will show. A toast message will be displayed if the passwords entered are not matching, a different toast message will be displayed when successful in the signup.